Add confirmOrder

+++ app/src/main/java/com/hfad/coffeepos/domain/entity/Dish.kt
@@ -2,6 +2,6 @@ package com.hfad.coffeepos.domain.entity
 
 data class Dish(
     val name: String? = null,
-    val cost: Float? = null,
-    val ingredients: HashMap<String?, Float?>? = null
+    val cost: Double? = null,
+    val ingredients: HashMap<String?, Double?>? = null
 )
