+++ app/src/main/java/com/hfad/coffeepos/domain/entity/Ingredient.kt
@@ -2,7 +2,7 @@ package com.hfad.coffeepos.domain.entity
 
 data class Ingredient (
     val name : String? = null,
-    val cost : Float? = null,
-    val quantity : Float? = null,
+    val cost : Double? = null,
+    val quantity : Double? = null,
     val units : String? = null
     )
